pyproject.toml
[project]
name = "protrack"
version = "0.1.0"
description = ""
authors = [
    {name = "Victory Art"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.2,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "boto3 (>=1.42.78,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "argon2-cffi (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]
